Web/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
numpy==1.26.4
opencv-python-headless==4.9.0.80
insightface==0.7.3
onnxruntime==1.17.3
Pillow==10.3.0
